Rework slightly the storage of the list of fixpoints

+++ src/main/java/org/colomoto/biolqm/tool/fixpoints/FixpointList.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.colomoto.biolqm.LogicalModel;
 import org.colomoto.biolqm.NodeInfo;
 
-public class FixpointList extends ArrayList<int[]> {
+public class FixpointList extends ArrayList<byte[]> {
 
 	public final List<NodeInfo> nodes;
